@@ -6,7 +6,8 @@ using System.Security.Cryptography.X509Certificates;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using VpnHood.Common.Messaging;
 using VpnHood.Server;
-using VpnHood.Server.AccessServers;
+using VpnHood.Server.Providers.FileAccessServerProvider;
+using VpnHood.Server.Providers.HttpAccessServerProvider;
 
 namespace VpnHood.Test.Tests;
 
@@ -21,8 +22,8 @@ public class FileAccessServerTest
         var publicEndPoints = new[] { IPEndPoint.Parse("127.0.0.1:443") };
 
         // Create accessServer
-        using TestEmbedIoAccessServer testRestAccessServer = new(fileAccessServer);
-        var accessServer = new RestAccessServer(new RestAccessServerOptions(testRestAccessServer.BaseUri.AbsoluteUri, "Bearer xxx"));
+        using TestEmbedIoAccessServer testHttpAccessServer = new(fileAccessServer);
+        var accessServer = HttpAccessServer.Create(new HttpAccessServerOptions(testHttpAccessServer.BaseUri, "Bearer xxx"));
 
         // ************
         // *** TEST ***: default cert must be used when there is no InternalEndPoint
@@ -115,13 +116,13 @@ public class FileAccessServerTest
 
         // ************
         // *** TEST ***: add sent and receive bytes
-        var response = accessServer1.Session_AddUsage(sessionResponse.SessionId, false,
+        var response = accessServer1.Session_AddUsage(sessionResponse.SessionId, 
             new UsageInfo { SentTraffic = 20, ReceivedTraffic = 10 }).Result;
         Assert.AreEqual(SessionErrorCode.Ok, response.ErrorCode, response.ErrorMessage);
         Assert.AreEqual(20, response.AccessUsage?.SentTraffic);
         Assert.AreEqual(10, response.AccessUsage?.ReceivedTraffic);
 
-        response = accessServer1.Session_AddUsage(sessionResponse.SessionId, false,
+        response = accessServer1.Session_AddUsage(sessionResponse.SessionId, 
             new UsageInfo { SentTraffic = 20, ReceivedTraffic = 10 }).Result;
         Assert.AreEqual(SessionErrorCode.Ok, response.ErrorCode, response.ErrorMessage);
         Assert.AreEqual(40, response.AccessUsage?.SentTraffic);
@@ -134,7 +135,7 @@ public class FileAccessServerTest
         Assert.AreEqual(20, response.AccessUsage?.ReceivedTraffic);
 
         // close session
-        response = accessServer1.Session_AddUsage(sessionResponse.SessionId, true,
+        response = accessServer1.Session_Close(sessionResponse.SessionId, 
             new UsageInfo { SentTraffic = 20, ReceivedTraffic = 10 }).Result;
         Assert.AreEqual(SessionErrorCode.SessionClosed, response.ErrorCode, response.ErrorMessage);
         Assert.AreEqual(60, response.AccessUsage?.SentTraffic);
