@@ -7,7 +7,7 @@ internal class SimpleLogger : ILogger
 {
     private readonly LoggerExternalScopeProvider _scopeProvider = new();
 
-    public IDisposable BeginScope<TState>(TState state)
+    public IDisposable BeginScope<TState>(TState state) where TState : notnull
     {
         return _scopeProvider.Push(state);
     }
