@@ -11,7 +11,8 @@ public interface IAccessServer : IDisposable
     bool IsMaintenanceMode { get; }
     Task<SessionResponseEx> Session_Create(SessionRequestEx sessionRequestEx);
     Task<SessionResponseEx> Session_Get(uint sessionId, IPEndPoint hostEndPoint, IPAddress? clientIp);
-    Task<ResponseBase> Session_AddUsage(uint sessionId, bool closeSession, UsageInfo usageInfo);
+    Task<ResponseBase> Session_AddUsage(uint sessionId, UsageInfo usageInfo);
+    Task<ResponseBase> Session_Close(uint sessionId, UsageInfo usageInfo);
     Task<ServerCommand> Server_UpdateStatus(ServerStatus serverStatus);
     Task<ServerConfig> Server_Configure(ServerInfo serverInfo);
     Task<byte[]> GetSslCertificateData(IPEndPoint hostEndPoint);
