@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net;
 
-namespace VpnHood.Server.AccessServers;
+namespace VpnHood.Server.Providers.FileAccessServerProvider;
 
 public class FileAccessServerOptions : ServerConfig
 {
